Screen package names changed. First version of currency changing added.
