+++ app/src/main/java/com/example/financetracker/data/api/model/AccountUpdateRequestModel.kt
@@ -1,0 +1,14 @@
+package com.example.financetracker.data.api.model
+
+import com.google.gson.annotations.SerializedName
+
+data class AccountUpdateRequestModel(
+    @SerializedName("name")
+    val name: String,
+
+    @SerializedName("balance")
+    val balance: String,
+
+    @SerializedName("currency")
+    val currency: String
+)
