+++ app/src/main/java/com/example/financetracker/domain/FinanceRepository.kt
@@ -1,6 +1,8 @@
 package com.example.financetracker.domain
 
+import com.example.financetracker.data.api.model.AccountModel
 import com.example.financetracker.data.api.model.AccountResponseModel
+import com.example.financetracker.data.api.model.AccountUpdateRequestModel
 import com.example.financetracker.data.api.model.CategoryModel
 import com.example.financetracker.data.api.model.TransactionModel
 import com.example.financetracker.utils.Result
@@ -16,4 +18,6 @@ interface FinanceRepository {
     suspend fun getAccountById(): Result<AccountResponseModel>
 
     suspend fun getAllCategories(): Result<List<CategoryModel>>
+
+    suspend fun putAccountById(name: String, balance: String, currency: String): Result<AccountModel>
 }
