+++ app/src/main/java/com/example/financetracker/presentation/screens/Income/IncomeUIState.kt
@@ -1,10 +1,10 @@
-package com.example.financetracker.presentation.screens.ExpensesScreen
+package com.example.financetracker.presentation.screens.Income
 
 import com.example.financetracker.utils.Result
 import com.example.financetracker.data.api.model.TransactionModel
 
 /**
- * UI state data class for the ExpensesScreen.
+ * UI state data class for the IncomeScreen.
  *
  * Holds the state of loading, list of transactions, and possible error.
  *
@@ -12,7 +12,7 @@ import com.example.financetracker.data.api.model.TransactionModel
  * @property transactions The list of fetched transactions.
  * @property error The occurred error during data loading (possibly).
  */
-data class ExpensesUIState(
+data class IncomeUIState(
     val isLoading: Boolean = false,
     val transactions: List<TransactionModel> = emptyList(),
     val error: Result.Error? = null,
