+++ app/src/main/java/com/example/financetracker/presentation/screens/expenses/ExpensesScreen.kt
@@ -1,56 +1,61 @@
-package com.example.financetracker.presentation.screens.MyAccountScreen
+package com.example.financetracker.presentation.screens.expenses
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.financetracker.R
+import com.example.financetracker.presentation.components.HandleErrors
 import com.example.financetracker.presentation.components.HorizontalItem
 import com.example.financetracker.presentation.components.PlusFloatingActionButton
 import com.example.financetracker.presentation.components.TopBar
-import com.example.financetracker.data.api.model.AccountBriefModel
 import com.example.financetracker.presentation.navigation.Screen
 import com.example.financetracker.ui.theme.Green
 import com.example.financetracker.ui.theme.LightGreen
-import com.example.financetracker.ui.theme.surface
 import com.example.financetracker.ui.theme.onSurface
-import com.example.financetracker.ui.theme.outlineVariant
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import com.example.financetracker.presentation.components.HandleErrors
+import com.example.financetracker.ui.theme.surface
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun MyAccountScreen(
+fun ExpensesScreen(
     onNavigateTo: (Screen) -> Unit,
-    viewModel: MyAccountViewModel = hiltViewModel()
+    viewModel: ExpensesViewModel = hiltViewModel(),
 ) {
 
-    val myAccountState by viewModel.accountState
+    val expensesState by viewModel.expensesState
 
+    val startOfDay = LocalDate.now().atStartOfDay()
+    val endOfDay = startOfDay.plusDays(1).minusSeconds(1)
+    val startDate = startOfDay.format(DateTimeFormatter.ISO_LOCAL_DATE)
+    val endDate = endOfDay.format(DateTimeFormatter.ISO_LOCAL_DATE)
     LaunchedEffect(Unit) {
-        viewModel.getAccountById()
+        viewModel.getAllExpenses(startDate, endDate)
     }
 
     val snackbarHostState = remember { SnackbarHostState() }
     HandleErrors(
-        error = myAccountState.error,
+        error = expensesState.error,
         onErrorHandled = { viewModel.clearError() },
         snackbarHostState = snackbarHostState
     )
@@ -66,40 +71,48 @@ fun MyAccountScreen(
                 .background(surface)
         ) {
             TopBar(
-                title = stringResource(R.string.myAccount_topbar),
-                rightIcon = R.drawable.ic_edit,
-                onRightIconClick = {},
+                title = stringResource(R.string.expenses_topbar),
+                rightIcon = R.drawable.ic_history,
+                onRightIconClick = { onNavigateTo(Screen.ExpensesHistory) },
                 colors = TopAppBarDefaults.topAppBarColors(
                     containerColor = Green,
                     titleContentColor = onSurface,
                 ),
             )
 
-            HorizontalItem(
-                modifier = Modifier
-                    .background(LightGreen)
-                    .height(56.dp),
-                title = stringResource(R.string.myAccount_balance),
-                emoji = "💰",
-                contentUpper = myAccountState.account?.balance,
-                icon =  R.drawable.ic_arrow_detail,
-            )
 
-            HorizontalDivider(
-                thickness = 1.dp,
-                color = outlineVariant
-            )
+            val expensesOnly = expensesState.transactions.filter { !it.category.isIncome }
+            val totalExpenses = expensesOnly.sumOf { transaction -> transaction.amount.toDoubleOrNull() ?: 0.0 }
+            val currency = expensesOnly.firstOrNull()?.account?.currency
+
+            val formattedTotal = "%,.2f %s".format(totalExpenses, currency)
 
             HorizontalItem(
                 modifier = Modifier
                     .background(LightGreen)
                     .height(56.dp),
-                title = stringResource(R.string.myAccount_currency),
-                contentUpper = myAccountState.account?.currency,
-                icon =  R.drawable.ic_arrow_detail,
+                title = stringResource(R.string.expenses_sum),
+                contentUpper = formattedTotal,
+                showDivider = true,
             )
+            
+            LazyColumn (
+                contentPadding = PaddingValues(bottom = 1.dp) // to show last divider
+            ){
+                items(expensesState.transactions.filter { !it.category.isIncome }) { item ->
+                    HorizontalItem(
+                        modifier = Modifier.height(70.dp),
+                        emoji = item.category.emoji,
+                        title = item.category.name,
+                        subtitle = item.comment,
+                        contentUpper = "${item.amount} ${item.account.currency}",
+                        icon = R.drawable.ic_arrow_detail,
+                        showDivider = true,
+                    )
+                }
+            }
 
-            if (myAccountState.isLoading) {
+            if (expensesState.isLoading) {
                 Box(
                     modifier = Modifier.fillMaxSize(),
                     contentAlignment = Alignment.Center,
@@ -110,7 +123,6 @@ fun MyAccountScreen(
                 }
             }
         }
-
 
         PlusFloatingActionButton(
             modifier = Modifier
