+++ app/src/main/java/com/example/financetracker/presentation/screens/expenses/ExpensesUIState.kt
@@ -1,10 +1,10 @@
-package com.example.financetracker.presentation.screens.IncomeScreen
+package com.example.financetracker.presentation.screens.expenses
 
 import com.example.financetracker.utils.Result
 import com.example.financetracker.data.api.model.TransactionModel
 
 /**
- * UI state data class for the IncomeScreen.
+ * UI state data class for the ExpensesScreen.
  *
  * Holds the state of loading, list of transactions, and possible error.
  *
@@ -12,7 +12,7 @@ import com.example.financetracker.data.api.model.TransactionModel
  * @property transactions The list of fetched transactions.
  * @property error The occurred error during data loading (possibly).
  */
-data class IncomeUIState(
+data class ExpensesUIState(
     val isLoading: Boolean = false,
     val transactions: List<TransactionModel> = emptyList(),
     val error: Result.Error? = null,
