+++ app/src/main/java/com/example/financetracker/presentation/screens/my_articles/MyArticlesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.financetracker.presentation.screens.ExpensesScreen
+package com.example.financetracker.presentation.screens.my_articles
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.CircularProgressIndicator
@@ -16,46 +15,40 @@ import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.financetracker.R
-import com.example.financetracker.presentation.components.HandleErrors
 import com.example.financetracker.presentation.components.HorizontalItem
-import com.example.financetracker.presentation.components.PlusFloatingActionButton
 import com.example.financetracker.presentation.components.TopBar
 import com.example.financetracker.presentation.navigation.Screen
 import com.example.financetracker.ui.theme.Green
-import com.example.financetracker.ui.theme.LightGreen
-import com.example.financetracker.ui.theme.onSurface
 import com.example.financetracker.ui.theme.surface
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
+import com.example.financetracker.ui.theme.onSurface
+import com.example.financetracker.ui.theme.onSurfaceVariant
+import com.example.financetracker.ui.theme.surfaceContainerHigh
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.res.stringResource
+import com.example.financetracker.presentation.components.HandleErrors
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ExpensesScreen(
+fun MyArticlesScreen(
     onNavigateTo: (Screen) -> Unit,
-    viewModel: ExpensesViewModel = hiltViewModel(),
+    viewModel: MyArticlesViewModel = hiltViewModel()
 ) {
 
-    val expensesState by viewModel.expensesState
+    val myArticlesState by viewModel.articlesState
 
-    val startOfDay = LocalDate.now().atStartOfDay()
-    val endOfDay = startOfDay.plusDays(1).minusSeconds(1)
-    val startDate = startOfDay.format(DateTimeFormatter.ISO_LOCAL_DATE)
-    val endDate = endOfDay.format(DateTimeFormatter.ISO_LOCAL_DATE)
     LaunchedEffect(Unit) {
-        viewModel.getAllExpenses(startDate, endDate)
+        viewModel.getAllArticles()
     }
 
     val snackbarHostState = remember { SnackbarHostState() }
     HandleErrors(
-        error = expensesState.error,
+        error = myArticlesState.error,
         onErrorHandled = { viewModel.clearError() },
         snackbarHostState = snackbarHostState
     )
@@ -71,48 +64,37 @@ fun ExpensesScreen(
                 .background(surface)
         ) {
             TopBar(
-                title = stringResource(R.string.expenses_topbar),
-                rightIcon = R.drawable.ic_history,
-                onRightIconClick = { onNavigateTo(Screen.ExpensesHistory) },
+                title = stringResource(R.string.myArticles_topbar),
                 colors = TopAppBarDefaults.topAppBarColors(
                     containerColor = Green,
                     titleContentColor = onSurface,
                 ),
             )
 
-
-            val expensesOnly = expensesState.transactions.filter { !it.category.isIncome }
-            val totalExpenses = expensesOnly.sumOf { transaction -> transaction.amount.toDoubleOrNull() ?: 0.0 }
-            val currency = expensesOnly.firstOrNull()?.account?.currency
-
-            val formattedTotal = "%,.2f %s".format(totalExpenses, currency)
-
             HorizontalItem(
                 modifier = Modifier
-                    .background(LightGreen)
+                    .background(surfaceContainerHigh)
                     .height(56.dp),
-                title = stringResource(R.string.expenses_sum),
-                contentUpper = formattedTotal,
+                title = stringResource(R.string.myArticles_search),
+                titleColor = onSurfaceVariant,
+                icon = R.drawable.ic_search,
                 showDivider = true,
             )
-            
+
             LazyColumn (
                 contentPadding = PaddingValues(bottom = 1.dp) // to show last divider
             ){
-                items(expensesState.transactions.filter { !it.category.isIncome }) { item ->
+                items(myArticlesState.categories) { item ->
                     HorizontalItem(
                         modifier = Modifier.height(70.dp),
-                        emoji = item.category.emoji,
-                        title = item.category.name,
-                        subtitle = item.comment,
-                        contentUpper = "${item.amount} ${item.account.currency}",
-                        icon = R.drawable.ic_arrow_detail,
-                        showDivider = true,
+                        emoji = item.emoji,
+                        title = item.name,
+                        showDivider = true
                     )
                 }
             }
 
-            if (expensesState.isLoading) {
+            if (myArticlesState.isLoading) {
                 Box(
                     modifier = Modifier.fillMaxSize(),
                     contentAlignment = Alignment.Center,
@@ -122,14 +104,6 @@ fun ExpensesScreen(
                     )
                 }
             }
-        }
-
-        PlusFloatingActionButton(
-            modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .padding(end = 16.dp, bottom = 16.dp),
-        ) {
-
         }
 
         // snackbar
