+++ app/src/main/java/com/example/financetracker/presentation/screens/my_articles/MyArticlesUIState.kt
@@ -1,4 +1,4 @@
-package com.example.financetracker.presentation.screens.MyArticlesScreen
+package com.example.financetracker.presentation.screens.my_articles
 
 import com.example.financetracker.utils.Result
 import com.example.financetracker.data.api.model.CategoryModel
