+++ app/src/main/res/values/strings.xml
@@ -27,4 +27,5 @@
     <string name="settings_syncronization">Синхронизация</string>
     <string name="settings_language">Язык</string>
     <string name="settings_about">О программе</string>
+    <string name="myAccount_accountName">Название счета</string>
 </resources>
